show conversation history in drawer on mobile device
